modules for generating a number, calculating bulls and cows were separated
